Fixed the parsing of numbers.
Added a "mixed tokens" test which verifies a long string of tokens without resetting the parser each time. It checks all the operators.
The exponent does not require a period: 3e3 is now valid, as is 3.e3.
We properly recognize 33.length as 33, a member, and a field.
The parser properly recognize a range such as 3..5, instead of seeing an error or two floats (3. and .5).
We accept the infinity character (0x221E) as the Infinity value.
Moved the match_expression() function to the correct location in the source.
Added the compare (<=>) operator in the parser.
Fixed a bug in the parsing of '/' which would eventually duplicate a character.
Added support for numeric types (use <name> as <min> .. <max>; or use <name> as mod <modulo>;).
Renamed the rotate operator using the % character ('<%' and '>%' instead of '!<' and '!>').
Use the C++ versions of the strtoll() and strtod() functions which accept std::string directly.
Updated the tests as required.
Various cleanups.