Added scripts plus cmake targets that build debian packages at LP.
This is a first pass at the functionality. Right now I put in two
targets per project (at the root CMakeLists.txt level): debuild and
dput.

So to build the source package for, say, cassview, you would type:
$ cd build
$ make cassview-debuild

Then to upload the package for building, you would:
$ make cassview-dput

At the root level, to build all source packages:
$ make debuild

Likewise, to upload those source packages:
$ make dput

Now I did tie the two targets together--because I wanted to be able to
generate all of the source packages, then upload them. Also, the
package generation bumps the version in the debian/changelog file and
writes it out, for each project. So I wanted that as a stand-alone step.
Yes this means you can goof and run the dput target, but at least dput
won't re-upload a package that has a file with a .ppa.upload extension
that matches your package. Or if the .changes file doesn't exist, the
dput target just fails with an error.

Two shell scripts make this magic work: SnapBuildMakeSourcePackage.sh
and SnapBuildDPutPackage.sh.

TODO: the package generator should be smart enough to modify the
debian/control file and update the version number for any dependency
that will be generated that we rely on. Otherwise LP will just build
them all in parallel, instead of stopping with "Dependency Wait."