Fixed several bugs in the JSON and the rc implementations.
Added a function to set the value of a JSON otherwise we cannot create one from memory.
Added a test for cyclical structures in JSON since we cannot save those.
Added a call to the As2JsRCUnitTests::test_empty_home test to include that in the coverage.
Added a StringOutput so we can write tests just using in-memory streams.
Fixed the JSON database header to use // instead of # for comments.
Added tests to verify that the set/get functions of the JSON value are all used with valid data.
JSON Verify indexes in case they are out of bounds.
Switch a certain number of std::stringstream to instead use the std::to_string() function;
Fixed the read_js_value() function so it actual works as expected.
Fixed some error position references in JSON, using the node that generates the error instead of the current position.
Fixed the JSON reader so it properly checks for commas between items in an object or array.
Fixed the JSON parser to correctly support negative numbers.
Correctly check whether a value is defined in the output() function, if not, throw.
Made the output header optional.
Fixed a bug in the rc.cpp implementation, it would print a pointer instead of the directory for directories starting with ~.
Fixed the position used while reading rc files.
Fixed the message tests as it was not properly reseting its warning and error counts.
Added a test to verify the StringOutput stream type.
Various clean ups.