Test the Database class.
Fixed the Database::Element::set_line() which would save the line as a boolean instead of a number.
Let the parsing of a package definition continue even if some entries are not objects as expected.
Create an f_value if the database is missing so calling load() again return true again.
Properly set the JSON value when adding packages.
Fixed the match_pattern() function which would loop through name but test the pattern variable.
Fixed the f_line definition in the database Element class so it get properly initialized to 1.
Now output a BOM at the start of the JSON files to make sure it gets read as utf-8.
Various clean ups.