Enhanced the popup and editor for the case when a popup is opened with an editor in it using an IFRAME.
Added the ability to mark an editor form as saved ("ignore changes").
Fixed the SAVE_MODE_... making them static (and public in the class declaration).
Added an editor earlyClose function so you can get a popup per editor form being closed at any time.
Added a shadow to the default popup so if opened over another window without the darken feature, it still detaches itself from the background.
Added a structure of the PopupData to simplify the other definitions.
Added a variable to accept the possibility that a popup should not have a close button.
Fixed the click() function on the popup close button. It was actually being re-added each time the popup was opened!
Fixed the calling of the show() and hide() callbacks, now those are called AFTER the animation is done.
Added the capability to be called when a popup is about to be closed (beforeHide() callback.)
Added a messageBox() function to avoid using things like the confirm() or alert() functions.