Fixed a bug in the parser which would eventually use a null node pointer.
Fixed the program loop which would stop on 'not EOF' instead of 'EOF'.
Added a coverage option so one can write coverage tests. Nothing implemented in that regard yet though.
Updated the lexer test to include the new coverage option.
Changed the debug lexer with a message so we can intercept it in our tests.
Added to the parser all the missing node types now supported as attributes.
Fixed the JSON test to properly handle the BOM added by the save() function.
Various clean ups.