As I am starting to seriously test the parser, getting quite a few bug fixed!
Fixed the return directive which had a reversed test.
Fixed the test for the case a function return type is set to "Never", it would test so only if the node was of type VOID.
Fixed several f_node references that were used instead of the 'node' or 'param' or some other variable.
Fixed the contract that would access the null node parameter.
Some clean ups and error message enhancements.