Progress on the parser. The function is 100% covered, although the coverage test shows 97.2% instead... to bad!
Made the NODE_NAME accept a string so we can use them for objects (instead of NOTE_TYPE which should not support strings.)
Fixed the node test accordingly.
Fixed the list of directives in the parsing of a 'class'. It would just create an empty list.
The skip on a missing ';' or other such error in the list of directives stops on the '{' token too now.
Moved the value of object literals inside a NODE_SET.
Fixed the has_out usage in the list of parameters: it would prevent any '...' when only the '...' cannot be an out.
Moved the function type under a NODE_TYPE.
Moved the parameter type under a NODE_TYPE.
Allow the parsing of the assignment for a '...' even though it does not accept it.
Revisit some of the error messages to strengthen them.
Fixed the variable check of a comma, it was inverted.
Moved the very large JSON code of the parser test to a separate .json file.
Added a "tool" to convert the JSON code to a C++ string so we can write it in JSON, and still have it as a string in the test.
Added parser tests to make sure flags are set as expected.
Added parser tests to make sure attributes are set as expected.
Strengthen the use of the verbose flag in the JSON data.
Properly test the line number on which errors occur.
Renamed the parser test as "test_parser" instead of "test_basics" since there is just one testing everything.
Adding the native scripts that I wrote a while back and had not yet added. It's not even remotely tested.