The package file is covered 100% and has a few other tests too. The 'enum' parsing is 100% covered too.
Updated the coverage test shell script to work with a --full option and generate HTML for the lcov website.
The coverage script uses 'bash' now instead of 'sh' to check pipe results.
Added necessary code to test the links of a node. Really in the parser we only get link attributes in the directive list.
Added support to accept '..' and '...' in lists of package names.
Fixed a possible loop in parsing a package name, we skip strings found in the list of package names.
Added an error in case two commas appear in a row in a list of options after an import.
Added an error for the import implementation.
Added a CLASS flag for enumerations.
Fixed the implementation of the namespace block, it now includes an actual block!
A namespace without a name is acceptable and automatically marked as private.
Added an error in case someone creates an unnamed "forward" enumeration.
Fixed the bug of not saving the identifier string in the variable.
Fixed three error messages as they were marked as warnings.
Allow the enumeration loop to end even if identifiers are missing (it could loop forever in the previous version.)
Allow a semicolon to stop an enumeration definition even if '}' is missing.
Fixed the add 1 where the 1 had to be a valid Int64 value.
Fixed a bug where I was testing NODE_NULL instead of NODE_COMMA.
Fixed the json_to_string tool so it deletes the output file if the parser fails en route.
Fixed a couple of problems with the string test: very small doubles are output with the scientific notation at times (the 'e' in the string) and true/false tests are checking exactly 0.0 and not 0.0.