100% coverage of a few more parser files. Still have to finish up the expression and directive.
Transformed the extended operators so they generate a "not allowed" error when used in the parser, instead of not generating them in the lexer. This makes it a huge whole lot easier to test/debug and thus it makes a lot more sense to do it that way.
Transformed the for() loop to include explicit CONST and IN flags.
Removed the DEBUG_LEXER option. It is not needed now that we have full coverage tests, we can simply write a test to make sure we get the right data.
Removed the TRACE_TO_OBJECT option, that was transforming the trace() function call to an object named Trace. We only will have an object in as2js.
Added an option named UNSAFE_MATH to allow mathematical expression simplifications that are not correct with floating points (i.e. a * b + a * c !== a * (b + c), at least no with floating points.)
Started using values other than just 0 and 1. We support 2 and 3 for two other options (extended operators and statements.)
Gave the option_value_t an explicit size instead of using 'long'.
Added a mechanism to err in case the if(),while(),for()... are used without an explicit block ('{ ... }').
Added a mechanism to err on C/C++ assignments ('=') and instead require the Pascal/Ada assignment (':=').
Explicitly prevent append_child() to be called with a NULL pointer.
The STRICT option is now cleared by default and not always forced to 1. Many things are probably not going to be strict by default and thus it did not really make sense to force the flag.
Documented the options functions.
If parsing an expression does not include any valid primary expression, then we still create a node (FALSE) because expression() cannot return NULL.
Simplified the extends and implements as expression() cannot return a NULL pointer.
Fixed many errors (removed contractions, added period, changed " with ', message when not appropriate--copy/paste mistake...)
Properly test attributes used without a valid declaration and return an error instead of throwing.
Fixed the case of a named '...' parameter.
Fixed the multiplicative expression so it properly makes use of the match expression in both cases (left and right hand sides) instead of just the first case (left hand side.)
Actually implemented the parsing of the '!~' operator in the parser expression.
Fixed the handling of a scoped dynamic identifier because the identifier would not get marked as variable (dynamic). Also check for the closing parenthesis and remove it if present.
Better handle a badly formed object literal: i.e. continue to parse as an object after a missing comma, unless a semi-colon is found.
Added support for catch(e if <expr>) as Mozilla supports.
Fixed a test in the pragma that used '==' instead of '!='.
Force a comma between pragmas to better support potential extensions. (there is no real definition of pragmas in ActionScript and ECMAScript.)
Fixed the fact that if the parser is created without an option object it would not be capable of changing options (so the 'use <pragma name>' would do nothing.)
Properly check pragma values as values instead of just a true/false flag (it works properly now with the prima syntax: use strict(3)?)
The case directive now always recognize the range syntax (case -4 ... 4: ...) but generates an error if extended statements are not allowed.
Properly detect that catch(void) was used (it is not legal).
Fixed up the for() directive to detect the IN in variable declaration at the right location.
Fixed the FOREACH test so it works properly.
Fixed the goto directive so we actually get the identifier string in it.
Better handle the case of an 'if' statement missing the 'if' block statement and being followed by an 'else'.
Fixed the use of with() following the switch() directive so unknown operators get skipped (otherwise we get many additional errors.)
Fixed the switch directive so if no attributes are specified we do not crash.
Allow rethrowing by using the 'throw;' statement without an expression like in C++.
Completely prevent the 'with' instruction when in strict mode (NOT the with following a switch though.)
Actually moved the type of a variable declaration to a TYPE node.
Updated tests accordingly.