Got the parser_directive.cpp file 100% covered. Fixed a couple of bugs at the same time.
Added a to_identifier() function to the node to convert PRIVATE, PROTECTED, and PUBLIC.
Fixed the handling of a lone colon.
Fixed the handling of a colon preceeded by an identifier, private, protected, or public.
Fixed the fact that the directive() function would return without an error if EOF was found after a list of attributes.