Got the parser_expression.cpp file 100% covered. Fixed a couple of bugs at the same time.
Fixed the handling of an empty entry in an array (last entry).
Fixed the handling of the rest (...) in case it is the only argument.
Fixed the handling of a private, protected, and public label in an object literal.
Fixed a bug testing for a COLON when we wanted to test a for a COMMA.
Finally enhanced the parser test so it tells me the error name expected and the integer/float when not equal.
Cleaned up a few things (messages, use of break instead of return...)
Show the date on each step of the coverage script.