Wrong new optimizer implementation using tables so it can very quickly grow.
Fixed the doxy file to include .ci files (include files).
Added a mechanism to compare all the flags and attributes of a node.
Added the != operator to our String implemention.
Added support to convert strings to floats and integers.
Fixed floating point output so tailing zeroes and periods get removed.
Removed the options from the optimizer, all options that have an effect must already have been applied or been saved in the tree of nodes.
Documented the lexer implementation.
Documented the JSON implementation.
Added write support for NaN, Infinity, -Infinity to JSON (not in the spec.) The reader was already reading those as expected.
Added support for positive numbers with a + sign (+3 is now recognized.)
Updated the tests accordingly.