Changed the status handling from one single number to a pair of numbers (still fits in a uint32_t) allowing working processes to mark a page without changing its main status.
Updated and tested the documentation of the status_t class.
Removed the cached status from the path_info_t class because it could otherwise change under our feet.
Changed the code to support the new status_t class.
Changes so as to Save the status first, then the ownership. This makes a lot more sense, otherwise we end up with a page without a status for a few milliseconds.
Fixed the updating of pages so once the update is over we can revert back to the state from before the update.
Now mark pages being updated as such (working_t::UPDATING.)
Added a function to simplify the addition of error messages from JavaScript code (i.e. without having to each time create a DOM document.)
Fixed some function documentation so Doxygen pick them up properly.