Removed the owner from the branch and revision names.
Added a basic status to all pages so we can better handle removals.
Fixed a bug in the date parser which simplifies strings so the US date parsing would not be selected as expected generating problems with the default and other layouts.
Added a test of the return value from the string_to_date() function in the layout so it throws if we get -1.
Test the page status so the path plugin can return 404 if the page is marked as CREATE, DELETED or an unknown state.
Similarly, the page status MOVED makes the path plugin redirect you to the new location of the page.
Added an UNKNOWN_WORKING to be able to distinguish between 'do not modify working' and set the working state to 'not working'.
Added a couple of cases that were not taken in account of state change.
Added an RAII class to handle setting / resetting a page status.
Put pages receiving a working status in the "processing" table so we can eventually fix the status if the process fails, somehow.
Added a backend process in the content plugin to fix the working status if it is not NOT_WORKING for more than 10 minutes.
Made some progress on the clone function, although at this point the rows do not get copied properly. Working on that I discovered another bug which will be fixed in the next load.
Fixed a bug with the layout last updated date. However, I put a TODO because the current fix is not 100% correct.