Finished the clone implementation. This time links get updated appropriately.
Links get fixed by calling a callback on each plugin that has a link that may need tweaking. If the callback is not defined, the link is lost.
The clone process includes cloning all the children of a page.
Broke up the copy process differently as branches and revisions cannot be separated as they were and the row predicate cannot be used.
Save a lot more information (wow!) about what gets copied.
Prevent cloning (move/trash) of pages that are marked as "prevent delete" (mainly all the administrative pages.)
Added the unicity to the link data because when duplicating a link it is important to have the right unicity!
Fixed the SNAP_NAME_CONTENT_OWNER with SNAP_NAME_CONTENT_PRIMARY_OWNER. I removed the unused name.
Changed the page type links to have "page" on one side and "page_type" on the other. This makes it cleaner when a page represents itself.
The create_link() function will actually generate an error on the old usage of the "page_type". This will prevent new similar "errors".
Replicated that new page type change in the editor plugin.
Fixed the dbutils so it checks "branch" instead of "data" as a table name.
Fixed the last revision key because it needs to have a locale if present.
Added a log trace to all the backend (not actions.)
Fixed the processing backend which could end up with an invalid URI that it needs to ignore.
Fixed the links row_key QString() change as it was referencing the wrong arguments.
Create a get_name() function for the robots.txt plugin.
The shorturl plugin can now be avoided on specific pages by implementing the allow_shorturl signal (i.e. sitemap.xml is not part of the shorturl scheme anymore, for example).
Fixed a bug in the users linking an author to its page. The branch for the authre was wrong.
When creating a user, the content::created date was not created.
Properly delete the "processing" and "revision" tables in the snapdb tool.
Many other changes, document improvements, optimizations, etc.