Added an option to remove the long session feature.
Added a way to forbid a page from being used as the referrer for a redirect (for login and other similar boxes.)
Added support for a class attribute in the item of a dropdown widget.
Added a reset function to reset a dropdown widget to the system default.
Added a hide() and a show() to the EditorWidget class.
Fixed the search for an item in a dropdown as we want to first search for the value if defined, then the HTML.
Verify that the referrer is a valid page on our site (just in case...)
Created a new command line tool (snapsiteinfo) so we can peek and poke in the database at will.