Fixed a few problems in the editor.
Fixed the status, the update working state would overwrite the state.
Added a case to fix a node in status CREATE/CREATING that did not get set to NORMAL/NOT_WORKING.
Fixed the state test in the text and line edit widgets (was accessing the attribute instead of the tag).
Added a focus() function to the EditorWidget class.
Fixed the keydown() test so it actually works. Also allows the tab and enter keys.
Fixed the Instance of each JS object. (a static variable)