Added a new versions plugin to get the version of the libraries, plugins, tools used by snap.
Made the map of plugins available to the outside.
Added necessary functions to define the version of a plugin within the plugin too (instead of just the factory.)
Fixed the name of the "server" plugin (and filename) which was empty before.
Added a backend action to create and delete links.
A few enhancements.