Added optimizations for multiplicative operations.
Fixed the naming convention of the g_optimizer_optimize_... tables.
Fixed the order of the optimization tables, it is now alphabetical.
Fixed the order of the functions, a couple were misplaced.
Enhanced the floating point tests to allow a small elipson.
We test infinity separately and signed in the optimizer test!