Added a copy_branch_cells() function to copy branch data when a new branch is created.
Made the get_new_revision() copy the data even between branches, used when creating the first user branch.
Better handles the previous_revision variable by initializing it with the current revision instead of 1.
Added some missing entries pages (secure page and attachment).
Fixed a few documentation entries.
Fixed the save on the editor where a force_branch() was missing.
Started adding a radio button as an editor widget.
Added a makeadministrator action so we can setup a root and an administrator of a website.