Fixed problems with links from multiple branches (i.e. the same child has many "children" links: one per branch). The cell name had to include the branch.
Started a bit on having the tabindex attribute set to -1 when opening the popup window.
Removed the e.keyCode access since jQuery already fixes that one.
Added a throw to clearly distinguish (in the message) whether a ':' is found in the wrong place in a link name.
Removed checks of Cassandra table pointers, if not set, the problem is detected early anyway.
When deleting a link, test that the count is at least 10.
Updated next_link() and cleaned it up some.
Fixed some documentation still referencing "data" when we now use "branch".
Fixed some code to make use of the link info cell_name() and other such functions to avoid code duplication.
Added an outline for the close-popup button when focused (only at this point it cannot get focus.)
Fixed the initQsParams_ function so it gets 'search' in the 'location' and not directly in 'document.'
Added a function to transform anchors into buttons (i.e. to react to Spacebar when focused.)
Allow a widget to be marked as "popup-auto-focus" in a popup so it automatically gets focused when opened.