Added the Radio buttons functionality to support various selections in websites.
Added a test page to the editor to check all the widgets in one place.
Fixed the permission "administer/info" with "administer/website/info" as it was (possibly) supposed to be.
Fixed a bug loading editor widgets: the QDomDocument used with a name makes it a non Null DOM object!
Allow loading of widgets from Qt resources.
Allow loading of layout XML files from Qt resources. This one is still experimental. There is a problem with the layout name and loading from Qt resources.
Generate a warning if no widgets can be found for the editor. Without an error, we waste a lot more time finding bugs such as mispelled entries.
Added CSS for widgets in default Snap! editor forms. Some things were missing because of the way the widgets are used in a default page.
Fixed up the Layout CMakeLists.txt, a couple of dependencies were missing.