Mainly updated the documentation of the Node class. Also found a few bugs and made some enhancements.
Made the statistics and log header available on a local run (non published run) of dev/coverage.
Added a test in the Node class destructor to catch cases where f_lock is not zero.
Updated the Unicode characters to include new valid characters in identifiers.
Made many updates to the Node documentation to be closer to what it looks like today.
Fixed the index name to the get_param_depth() to 'j'.
Similarly fixed the index name to the get_param_index() to 'idx' (it returns 'j').
Fixed the set_param_index() which now verifies that 'j' is within the limit of the number of parameters.
Fixed the insert_child() and set_child() and replace_with() ... which had to verify that 'child' was not a null pointer.
Added an optimization for '~~~a' which is equivalent to '~a'.
Fixed a bug which would not properly see '!~a' as '!' then '~' then 'a'.
Extract the variable and function tests out of the large test_as2js_parser.json, added a couple test to each.
Updated the tests accordingly.