Added code to better handle case when the "sites" table does not exist (i.e. initialization without snapsetup).
Fixed the backends that loop: on a STOP signal, they now properly call exit(0) instead of returning (returning had the same effect as long as we had ONE website defined.)
Ameliorated the gathering of the UDP signal name by adding a function instead of hard coding it in multiple places, also moved them to SNAP_NAME_... declarations.
Added a few more fields to dbutils.
Added a flag to mark when a site is ready (one flag per site).