Added lib/mounts.cpp/h to read the list of mounted partitions on a Linux system.
Add process_list to find the list of currently running processes (uses procps library).
Fixed up the log implementation so that, until the log4cplus is setup, it prints out error in the console if the user is in a tty.
Updated the plugins, server, snap_child implementations to support snapwatchdog as its own server.
Allows the list of plugins to be defined in the configuration file in a way that prevents user defined additions in the website 'sites' table.
the handling of the command line first reads the configuration file so we can display the name which is now even more dynamic to the user.
Added snapwatchdog to the snapsignal, only that's really broken, we need to find a better way to handle such additions... (external .conf files)
Added a few missing tests in the snapwebsites CMakeLists.txt: Boost, LibXml2, and gjslint tool.
Added a new test to snapwebsites CMakeLists.txt: ProcPS
Various clean ups (added const, documentation, comments...)