Many enhancements to the permission plugin.
Pages can now have a "permissions::direct::action::..." field to define a specific permission on that page (before we could not distinguish between page permissions and group/type permissions).
Permission links now use "permissions::link_back" on one side so no side effects appear because of such.
Use get_name() to define all the permission field names. We may want to offer more (pre-defined) names though.
Enhanced the perms() snap_expr function to support specifying the login status (very important now!)
Enhanced the SHOW_RIGHTS output.
Made the improve_signature() event available to plugins.
Fixed the set_site_parameter() so it creates the "sites" table using the server create_table() function and wait for the table to be synchronized in the Cassandra cluster.
Simplified a few things in the die() function.
Changed the "invalid number of parameters" errors of the snap_expr to better exceptions when they do not represent an error in link with the number of parameters.
Changed the create_table() function to make sure tables get synchronized, even with a single node, I was getting errors because of that lately.
Allow for errors of attachments to return data that represent the error in the format of the attachment instead of an HTML error, already works with JS, CSS, and XML documents.
Added the "serverstats" to delete it when --drop-all-table is used with snapdb.
Removed double deletion of the "revisions" table.
Fixed 3 bugs in the deletelink action, 3 tests were opposites.
Fixed documentation and added many TODO/TBD/XXX entries to keep track of things to verify at a later time.