Fixed many small things so coverage is as close to 100% for everything that is expected to be done.
Fixed the optimizer definition in the compiler: it is now a static function anyone can call (no optimizer object because all options must be part of the tree.)
Added support and test of character FFFD as a representation of the value NaN as in Java.
Added optimizations for the <? and >? operators with literals.
Fixed the optimization of the `a * NaN` type of expressions: the input does not need to be an identifier, only to not have any side effects.
Test and fixed the simplified() function which did not accept signed numbers.
Completed test of the smart match operator.
Closely test the nearly_equal() function.
Cleaner up the optimizer.cpp file (i.e. all the old code is gone.)
Commented out various parts of the code to avoid low coverage results of things that are not likely going to ever be used.
Enhanced the coverage script so I can run one set of test to verify that each set properly covers what it is expected to cover by itself.
Changed the lcov call to remove the system include data.