Added necessary code to initialize a website (library object with background thread!).
Fixed the links BACK "pointers", they need to be specialized enough (i.e. view, edt group...)
Fixed the table initialization process: we have to get the pointer a second time after a create, so caching the pointer on the first call causes a problem when initializing those tables the very first time.
Fixed the sessions and users plugins: they now creates their table in the initial_update() function as expected.
Re-implement the #INIT=... command so it works like #START=... except for what is being sent back to the connection on errors and at the end.
Added Cassandra initialization information for the context so we support simple, local, and network and replication per data center.
Pass all the parameters to the snap_cassandra object so we can check them out at any point in time.
Fixed a couple of auto_ptr<>() with unique_ptr<>() as it should have been.
Enhanced the tcp_client so we can get our IP and port (from the socket).
Added a shared pointer definition for the tcp_client/tcp_server classes.
Added a synchornization call in the snapdb after dropping tables. This helps making sure we do not quit too soon, but we still get some timeouts when trying to add the layouts right after...
Enhanced documentation.