Added a window to handle the initialization of a website.
Made the getChilde() template available in the header file.
Fixed the problem with the snapmanager "crashing" (throwing) when a connection to the server fails.
Changed a critical error in a warning, if the lock table is not present that's not the end of the world.
Removed the #include to a boost file since we do not use boost at all and do not want the dependency.
Cleaned up some code here and there.