Added a signal so plugins can know when a new attachment was added and change things such as permissions.
Added support for an attachment identification.
Attempt a fix to the attachment filename. At this point that still does not work. But I think I know what's wrong.
Fixed the darkenPage() call when not redirecting to use the new_status and thus show the waiting effect when possible.
Various clean ups, code and documentation.