Got the very first Compiler test to work (fail) as expected. Fixed a few things in the process which right now breaks other tests which I will be fixing now.
Started change to support ONE options pointer for all the objects. Parser cannot allocate an Options object without sharing the exact same object with the compiler and the lexer.
Removed unused function isspace() from compiler.
Fixed a validty test when compiling the attributes: we wanted to check functions assigned the native keyword, nothing else.
Changed the Compiler::compile() function to return the number of errors the compiler generated. Not the total number of errors.
Only initialize the scope node if it has a chance to get used.
Compiler properly tests whether the g_db was already loaded.
Changed the lexer constructor to forcibly request for an input pointer and an options pointer. If either is missing, the function throws.
Removed the other checks on f_options since it cannot be NULL anymore.
Removed all the optimizer.cpp dead code.
Added support for class definitions to understand ': public ...' (and protected/private) so we can continue to parse properly, although we generate an error saying that the user needs to use 'extends' instead.
Got all the native script definitions to compile properly (fixed extends, Function, :=, !~, const, and a few other things).
Extracted the class tests of the parser to its own JSON file and added tests for the ': public ...' types of definitions.
Various clean ups and enhancements to the tests.