Added a new function to convert an attribute number to a string.
Enhanced the error message if an invalid attribute is found: show the name of the node and attribute.
Fixed the class so an empty class can be compiled as expected.
Fixed the attribute recursivity, if parent is PACKAGE or PROGRAM, stop the recursion.
Updated the coverage script to run the compiler tests.
Updated the node test to include the new attribute to string function.
Updated the compiler test to work with the new code.