Updated the lexer and parser tests so they are back to 100% coverage.
Fixed the coverage script when used with a specific test.
Added checks against 'extends' and 'implements' to make sure they are used as expected (legally).
Removed the f_options tests as now the pointer cannot legally be null.
Strengthen the class inheritance test in the parser.
Added proper support for an empty enumeration definition (although probably not useful!)
Broken up the parser JSON in many files, although it is not finished yet.
Added a few tests to the compiler.
Added tests to complete the lexer coverage: we now need to test the null pointers separately.