Added support for an editor timezone widget in the locale plugin.
Added a few funtions to fully support editor form/widget extensions.
Fixed the editor widget type initialization by adding the preInitializeWidget() function, called as the initializeWidget() function before. And now the initializeWidget() function is called after ALL widgets were initialized.
Verify that each EditorWidget object has a widgetContent object (jQuery length != 0) otherwise we've got a big problem in our XSLT output.
Added functions to Enable/Disable editor widgets.
Added a function so a user can retrieve the widget type.
Implemented the save button which worked exactly as expected in my one test...
Fixed a problem with focus and read-only: toolbar should not be shown in that case because it is anyway useless.
Fixed calls to itemClicked() which was using 'widget' instead of 'editor_widget' (we avoid a cast that way).
Added a -D to force explicit definitions of UnicodeStrings (ICU library).
Added a test against gettimeofday() in snap_child since it can return an error.
Added the set_timezone() and restore_timezone() functions in snap_child() to handle TZ appropriately.
Added a snap_io_exception exception.
Added an action named 'forceresetstatus' so an administrator can quickly reset the status of pages without having to run the backend.
Fixed the name of the editor.css attachment file (in the attribute).
Changed the editor-form.xsl by moving the <widget> tag outside of the sub-template. This allows for easy implementation of composite widgets.
Fixed a couple of problems with the <xsl:text> in QtPatterns. It simply doesn't work.
Added a few tags to the simplified-xslt20.dtd.
Started work on the feed plugin. Now that the list works like a charm, it will be easy to complete.
Fixed the no_shorturl link which was to robotstxt instead of the expected shorturl tag!
Applied a fix to avoid a crash when specifying an incorrect path to snapdb while decoding the files table (i.e. expecting an md5...)
Fixed various small bugs I caught as I was working on the new features.
Fixed some documentation in various places.
Updated the js-editor.odg document and corresponding PNG with information of the current implementation.