Reconfigured the locale plugin by breaking it in 3 so we have proper dependencies (no loops).
Finally changed the naming convention of the XXX-parser.xsl files, the "-parser" part must appear in the file declarations (just like the "-page" appears there.) There is only one exception at this point which we'll try to get rid off later.
Added the 3 locale plugins to the default .conf file.
Updated the set_timezone() and added a set_locale() in the snap_child class. However, these are really low level and other plugins should only make use of the locale plugin. I may retire those functions soon...
Fixed a bug in the signal "done" macro which was missing a '}'.
Reset the locale and timezone in the server on startup so that way we know about the exact behavior (at least we can hope so...)
Updated the dependencies.txt and sorted many existing entries.
Fixed the test of a default value for a dropdown, it was missing the possible default tag.
Fixed a bug in the handling of the default value of a timezone which would not remove the '_' character.
Fixed the auto-focus, with the changes to the initialization process the focus was broken as the wrong object .focus() function was being called.
Fixed the sitemapxml plugin generate_sitemapxml_impl() documentation which was referencing robotstxt. Also removed test on branch_table, no need.
Changed the filter [date] to make use of the locale plugin when the format is empty (default).
Send a signal once a user is marked as logged in wether from the login form or the cookie.
User now handles its own locale and timezone settings (since it can depend on the locale plugin without generating a loop.)
Removed the dependency (which was just a #include) to the sitemapxml from the users plugin.
Fixed various documentations.