Started work on the e-Commerce plugin.
Got an alpha version of the feed implementation. We still need to fix the dates and get a summary/teaser and settings (for icons, feed description, etc.)
Added a file to transform the feed data to an Atom feed.
Moved the feed settings under /admin/settings as expected by all plugins.
Added a type so we can mark feed files (attachments) as such.
Added the code to insert the feed links in the HTML header.
Added support so we can add a CDATA section in our feeds (QXmlPatterns does not offer that capability.)
Added support to allow us to add namespaces to tags defined in the feed output (QXmlPatterns does not offer that capability.)
Fixed several qdomxpath predicate tests which would throw instead of returning false.
XPath accepts Node Sets in predicates, if empty, return false.
If an attribute is not define, predicate returns an empty list instead of a list with a null node.
Fixed the "//" as we want to use the XPath descendants instead of just children.
Fixed usage of "*" in an XPath: it was not being skipped properly.
Fixed the cmd_save() function with a full fledge parser of the children path and attributes. This allows for attributes with any character other than mixed quotes.
Fixed some documentation in the editor.js, including closure compiler related markups.
Fixed one of the XPath in the form.cpp file.
Enhanced the cxpath-tests.sh script to check the newly fixed features.
Changed the html-header.xsl to use an 'href' instead of a 'format' for the alternate-format template. That way we can use that template for the feed links.
Moved the locale settings under /admin/settings as it ought to be.
Applied a couple small fixes to make the closure compiler happy about the locale-timezone.js file.